requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.2
numpy==1.23.5
pandas==1.5.1
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
yarl==1.8.1
